update trebuchet and trigger installer to accept wmf repos
this requires addon scripts, one to grab the right git repo for
each tool, and one to copy the appropriate files to the right
location on the appropriate host in the cluster (the master
for trebuchet and the host selected as the deployment server,
for trigger).

also updated the README to reflect these changes.

updated the install script (not yet tested) to patch only when
the versions of trigger/trebuchet don't have the specific patch.